Refactor project structure
- Besides merging constants modules and utils, I didn't touch the code. This is just modules reorganization to meet python standards.
- Updated uvicorn command in make and containerfile.
- Copy whole app in containerfile instead of picking individual files. As its build from repo, there should be no credentials there.
