+++ .gitignore
@@ -3,6 +3,7 @@ __pycache__
 .env*
 .vscode
 .DS_Store
+*.egg-info*
 
 # ignore venv
 venv
