--- app/constants.py
@@ -1,6 +1,0 @@
-# ols endpoint constants
-VALID = "VALID"
-INVALID = "INVALID"
-YAML = "YAML"
-NOYAML = "NOYAML"
-SOME_FAILURE = "some failure"
