+++ lightspeed_service/constants.py
@@ -1,3 +1,10 @@
+# ols endpoint constants
+VALID = "VALID"
+INVALID = "INVALID"
+YAML = "YAML"
+NOYAML = "NOYAML"
+SOME_FAILURE = "some failure"
+
 # templates
 SUMMARIZATION_TEMPLATE = """
 The following context contains several pieces of documentation. Please summarize the context for the user.
