+++ lightspeed_service/docs/docs_summarizer.py
@@ -1,11 +1,11 @@
 import os
 import llama_index
-import src.constants as constants
+from lightspeed_service import constants
 from dotenv import load_dotenv
 from llama_index import StorageContext, load_index_from_storage
 from llama_index.prompts import PromptTemplate
-from utils.model_context import get_watsonx_context
-from utils.logger import Logger
+from lightspeed_service.utils.model_context import get_watsonx_context
+from lightspeed_service.utils.logger import Logger
 
 load_dotenv()
