+++ lightspeed_service/indexer/indexer.py
@@ -3,8 +3,8 @@ from dotenv import load_dotenv
 import llama_index
 from llama_index import VectorStoreIndex, SimpleDirectoryReader
 from llama_index.storage.storage_context import StorageContext
-from utils.model_context import get_watsonx_context
-import src.constants as constants
+from lightspeed_service.utils.model_context import get_watsonx_context
+from lightspeed_service import constants
 
 llama_index.set_global_handler("simple")
